# back/root/build/swagger.yaml
swagger: "2.0"
info:
  description: "수화 교육 프로젝트 Swagger입니다."
  version: "1.0.0"
  title: "Sign Language Swagger"
host: "localhost:5000"
basePath: "/"
tags:
  - name: Sign Language
    description: "Sing Language in Alphabets"
  - name: Register User
    description: "Registering User"
  - name: Login User
    description: "Login User"
schemes:
  - "http"

paths:
  /register:
    post:
      tags:
        - Register User
      summary: "Register User"
      description: ""
      operationId: "Register User"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Register User"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
  /user:
    post:
      tags:
        - Login User
      summary: "Login User"
      description: ""
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Register User"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
    put:
      tags:
        - Login User
      summary: "Login User"
      description: ""
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: authorization
          in: header
          type: string
          required: true
        - in: "body"
          name: "body"
          description: "Register User"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
    patch:
      tags:
        - Login User
      summary: "Login User"
      description: ""
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: authorization
          in: header
          type: string
          required: true
        - in: "body"
          name: "body"
          description: "Register User"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"
    delete:
      tags:
        - Login User
      summary: "Login User"
      description: ""
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: authorization
          in: header
          type: string
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "405":
          description: "Invalid input"

  /hands:
    post:
      tags:
        - Sign Language
      description: "putting in Sign language image"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Register hand data"
          required: true
          schema:
            $ref: "#/definitions/Hand"
      responses:
        "201":
          description: "POST Successful."
        "405":
          description: "Invalid input"
    get:
      tags:
        - Sign Language
      description: "getting all alphabet dataset"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        "200":
          description: Return JSON

  /hands/{alphabet}:
    get:
      tags:
        - Sign Language
      description: "getting all alphabet dataset"
      consumes:
        - "applicaion/json"
      produces:
        - "applicaion/json"
      parameters:
        - in: path
          name: alphabet
          schema:
            type: string
          required: true
          description: alphabet string
      responses:
        "200":
          description: Return JSON

securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"

definitions:
  User:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
      username:
        type: "string"
    xml:
      name: "Order"
  Hand:
    type: object
    properties:
      alphabet:
        type: string
        example: "test"
      handVideo:
        type: string
        example: "test hand video"
      mouthVideo:
        type: string
        example: "test mouth video"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"